requirements.txt
amqp==2.5.2
argon2-cffi==19.2.0
asgiref==3.2.7
billiard==3.6.3.0
celery==4.4.2
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
defusedxml==0.6.0
diff-match-patch==20181111
Django==3.0.5
django-allauth==0.41.0
django-appconf==1.0.4
django-celery-beat==2.0.0
django-celery-email==3.0.0
django-ckeditor==5.9.0
django-crispy-forms==1.9.0
django-debug-toolbar==2.2
django-environ==0.4.5
django-extensions==2.2.9
django-formtools==2.2
django-import-export==2.0.2
django-js-asset==1.2.2
django-model-utils==4.0.0
django-redis==4.11.0
django-taggit==1.3.0
django-timezone-field==4.0
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
et-xmlfile==1.0.1
future==0.17.1
httplib2==0.17.2
idna==2.9
importlib-metadata==1.6.0
jdcal==1.4.1
kombu==4.6.8
MarkupPy==1.14
mysqlclient==1.4.6
oauthlib==3.1.0
odfpy==1.4.1
openpyxl==3.0.3
Pillow==7.1.1
pycparser==2.20
PyJWT==1.7.1
python-crontab==2.4.1
python-dateutil==2.8.1
python3-openid==3.1.0
pytz==2019.3
PyYAML==5.3.1
redis==3.4.1
requests==2.23.0
requests-oauthlib==1.3.0
six==1.12.0
sqlparse==0.3.1
tablib==1.1.0
urllib3==1.25.9
vine==1.3.0
xlrd==1.2.0
xlwt==1.3.0
zipp==3.1.0
